[test] Move dataset-free API tests into ctest
The image format and quaternion layout tests live in cuvslam_api_test, which is
declared with setup_app and so is never registered with ctest. CI runs ctest, so
these four tests have never actually executed anywhere despite needing no dataset.

Moves them to libs/cuvslam/test, where setup_test_project registers them, and
leaves cuvslam_api_test as what it really is: the edex replay suite that requires
--data_folder. All four pass.

The Odometry member of the image format fixture was named tracker, which now reads
as the unrelated public Tracker class; it is renamed to odometry.
